Add a tool for bumping workflow version
This uses xsltproc to edit the raw/info.plist.

The xslt approach is a little bit brittle and hacky in terms of the way
it uses an empty comment to stablise the xml emitted by Alfred, but i
think it's workable for now, and feels safer than non-xml based editing.

Also,
- Factor some of the git invocations more aggressively.
- Fix a couple of style improvements found by WebStorm