Add typesafety for buttons
Teach typescript how the `buttons` property in the details and the
`buttonReturned` property of the response for `displayDialog` are
linked.

You now get a typescript error if you mistype a button name when
checking `buttonReturned`.

I looked at the defaultAnswer / textReturned correspondence, and left a
comment in the code that it seems to complex for the ROI.